Added new userViewclass and Adminclass
