+++ Controller/AdminMenuController.cs
@@ -25,5 +25,30 @@
                 }
             }
         }
+
+        //public User FindUser(User l)
+        //{
+        //    {
+        //        var user = context.Users.FirstOrDefault(u => u.Name == loggedInUser.Name && u.Password == loggedInUser.Password);
+        //        if (user != null)
+        //        {
+        //            context.Users.Remove(user);
+        //            context.SaveChanges();
+        //            user = context.Users.FirstOrDefault(u => u.Name == loggedInUser.Name && u.Password == loggedInUser.Password);
+        //            if (user == null)
+        //            {
+        //                Console.WriteLine("You have succesfully deleted your account");
+        //                helper.PressEnter();
+        //            }
+        //            else
+        //            {
+        //                Console.WriteLine("Something went wrong with deleteing your account.");
+        //                MainMenuView menu = new();
+        //                menu.MainMenu(loggedInUser);
+        //            }
+        //        }
+        //    }
+        //    return null;
+        //}
     }
 }
